v1

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('email');
-            $table->integer('phone')->unique()->index();
+            $table->bigInteger('phone')->unique()->index();
             $table->string('language');
             $table->string('credentials');
             $table->string('profile');
